materia: materia_arbor scriptor - arbor -> STML (phasis I.3b)
Suite 5/5, assertiones CCXC, exitus 0. Porta shim adhuc CCCXLVI/CCCXLVI.

Scriptor arboris portatus: MCCCIX lineae, cum XL assertionibus propriis.
Lector nondum inceptus.

QUATTUOR MUTATA A SILVA, ET SUTURA UNA NOVA
  - genus lexematis per materia_lexicon solvitur; tabulae
    ORTHOGRAPHIAE et NOMINA_GENERUM non comitantur
  - 'valor portandus' = species==VERBATIM, non 'orthographia NIHIL
    et non EOF'
  - FORMA triviorum ex SPECIE venit: REPETITUM -> n=, TERMINATOR ->
    crlf, VERBATIM -> textus. Silva in quattuor genera nominata
    commutabat; lingua cuius spatium unum genus est (CSS) hic sine
    mutatione agit.
  - ancora et agnitio lexematis derivati per MateriaOrigoUncus

MateriaArborFrons sutura nova est, et DUOS uncos fert, non unum, causa
OCTETIM VISIBILI: attributa et liberi listae SEPARATAE sunt, et silva
'standard' ANTE 'f' scribit sed scissuras et originem POST <post>.
Uncus unus utrolibet fine ordinem utrumque reddere NON posset. Ergo
attributa_ornare prima, liberos_ornare ultima, attributa materiae in
medio.

LEX ANCORAE PORTATA CUM CAUTIONE SUA. silva_arbor.c:715-760 notat
lexemate solo adhibito CLXXVIII divergentias 'lexema/offset' super
corpus ortas - et probationem parvam id NON cepisse, quia 'int n = 0;'
lexema primum ad offset 0 sine triviis ducentibus habet, casus in quo
vitium evanescit. Ergo probatio hic AMBOS casus asserit, et vitium
plantatum id confirmavit: quaestione trivii ducentis ablata, tres
assertiones cadunt.

RUBRUM FALSUM, ET GEMINUM PEIUS EST

Post restitutionem plagulae bonae probatio RUBRA MANSIT. Fons octetim
idem ac bonus probatus, et adhuc cadebat. Causa:
    materia_arbor.c  21:27:07
    materia_arbor.o  21:27:07
Cursor '[ $src -nt $obj ]' rogabat. Mtimes secundo mensurantur, ergo
fons et obiectum EODEM SECUNDO scripta 'non novius' dant,
recompilatio TACITE omittitur, et suite contra obiectum prius currit.
Plagula parva, machina celeris - non rarum.

Geminum est falsi viridis quod silvam hodie momordit, ET PEIUS.
Viride falsum te peractum putare facit; RUBRUM FALSUM codici RECTO
diffidere facit, et motus proximus naturalis est 'emendare' quod
fractum numquam fuit.

Remedium: '! [ $obj -nt $src ]' - in aequalitate RECOMPILA. Verbum
unum, et aequalitas nunc in latus tutum cadit. Probatum plantando et
restituendo SINE reaedificatione coacta: III fracta, deinde XL
praeterita.

silva/compile_probationes.sh idem exemplar '-nt' fert et idem vitium.
NON tactum - silva gelata est et hoc Franis arbitrium - sed periculum
VIVUM ibi est, non theoreticum: idem cursor est qui hodie viride
falsum alia via peperit.

PROXIMUM: lector arboris (STML -> arbor), deinde shim ad documenta
STML conferenda extensus, deinde circuitus - quae porta phasis vera est.

+++ materia/probationes/probatio_materia_lexicon.c
@@ -92,7 +92,7 @@ s32 principale (vacuum)
         CREDO_AEQUALIS_S32 (iudicium.vitium, (s32)MATERIA_LEX_SANUM);
         CREDO_VERUM (ratum.ratum);
         CREDO_AEQUALIS_I32 (ratum.lexicon->numerus_generum,
-            (i32)LEXICON_C89_NUMERUS_GENERUM);
+            (s32)LEXICON_C89_NUMERUS_GENERUM);
     }
 
 
@@ -274,63 +274,63 @@ s32 principale (vacuum)
 
         /* Valores noti - erratum generationis grossum caperent */
         CREDO_VERUM (strcmp(materia_lexicon_titulus(&ratum,
-            (i32)LEXICON_C89_GENUS_IDENTIFICATOR), "IDENTIFICATOR") == ZEPHYRUM);
+            (s32)LEXICON_C89_GENUS_IDENTIFICATOR), "IDENTIFICATOR") == ZEPHYRUM);
         CREDO_VERUM (strcmp(materia_lexicon_orthographia(&ratum,
-            (i32)LEXICON_C89_GENUS_AUTO), "auto") == ZEPHYRUM);
+            (s32)LEXICON_C89_GENUS_AUTO), "auto") == ZEPHYRUM);
         CREDO_NIHIL (materia_lexicon_orthographia(&ratum,
-            (i32)LEXICON_C89_GENUS_IDENTIFICATOR));
+            (s32)LEXICON_C89_GENUS_IDENTIFICATOR));
         CREDO_VERUM (strcmp(materia_lexicon_orthographia(&ratum,
-            (i32)LEXICON_C89_GENUS_SPATIA), " ") == ZEPHYRUM);
+            (s32)LEXICON_C89_GENUS_SPATIA), " ") == ZEPHYRUM);
         CREDO_VERUM (strcmp(materia_lexicon_orthographia(&ratum,
-            (i32)LEXICON_C89_GENUS_NOVA_LINEA), "\n") == ZEPHYRUM);
+            (s32)LEXICON_C89_GENUS_NOVA_LINEA), "\n") == ZEPHYRUM);
         CREDO_VERUM (strcmp(materia_lexicon_orthographia(&ratum,
-            (i32)LEXICON_C89_GENUS_CONTINUATIO), "\\\n") == ZEPHYRUM);
+            (s32)LEXICON_C89_GENUS_CONTINUATIO), "\\\n") == ZEPHYRUM);
         CREDO_VERUM (strcmp(materia_lexicon_orthographia(&ratum,
-            (i32)LEXICON_C89_GENUS_EOF), "") == ZEPHYRUM);
+            (s32)LEXICON_C89_GENUS_EOF), "") == ZEPHYRUM);
 
         /* Munera propria */
         CREDO_AEQUALIS_S32 ((s32)materia_lexicon_munus(&ratum,
-            (i32)LEXICON_C89_GENUS_NOVA_LINEA), (s32)MATERIA_MUNUS_LINEA);
+            (s32)LEXICON_C89_GENUS_NOVA_LINEA), (s32)MATERIA_MUNUS_LINEA);
         CREDO_AEQUALIS_S32 ((s32)materia_lexicon_munus(&ratum,
-            (i32)LEXICON_C89_GENUS_CONTINUATIO), (s32)MATERIA_MUNUS_LAMINA);
+            (s32)LEXICON_C89_GENUS_CONTINUATIO), (s32)MATERIA_MUNUS_LAMINA);
         CREDO_AEQUALIS_S32 ((s32)materia_lexicon_munus(&ratum,
-            (i32)LEXICON_C89_GENUS_COMMENTUM_LINEA),
+            (s32)LEXICON_C89_GENUS_COMMENTUM_LINEA),
             (s32)MATERIA_MUNUS_COMMENTUM);
 
         /* textum_fert: VERBATIM sola. Commenta VERUM, spatia FALSUM -
          * id est ipsa divergentia a valor_portandus (quae utrisque
          * VERUM dabat) quam descriptor tollit. */
         CREDO_VERUM  (materia_lexicon_textum_fert(&ratum,
-            (i32)LEXICON_C89_GENUS_COMMENTUM_CLAUSUM));
+            (s32)LEXICON_C89_GENUS_COMMENTUM_CLAUSUM));
         CREDO_VERUM  (materia_lexicon_textum_fert(&ratum,
-            (i32)LEXICON_C89_GENUS_IDENTIFICATOR));
+            (s32)LEXICON_C89_GENUS_IDENTIFICATOR));
         CREDO_FALSUM (materia_lexicon_textum_fert(&ratum,
-            (i32)LEXICON_C89_GENUS_SPATIA));
+            (s32)LEXICON_C89_GENUS_SPATIA));
         CREDO_FALSUM (materia_lexicon_textum_fert(&ratum,
-            (i32)LEXICON_C89_GENUS_NOVA_LINEA));
+            (s32)LEXICON_C89_GENUS_NOVA_LINEA));
         CREDO_FALSUM (materia_lexicon_textum_fert(&ratum,
-            (i32)LEXICON_C89_GENUS_AUTO));
+            (s32)LEXICON_C89_GENUS_AUTO));
 
         /* trivium_est: commenta ET spatia, non identificator */
         CREDO_VERUM  (materia_lexicon_trivium_est(&ratum,
-            (i32)LEXICON_C89_GENUS_COMMENTUM_CLAUSUM));
+            (s32)LEXICON_C89_GENUS_COMMENTUM_CLAUSUM));
         CREDO_VERUM  (materia_lexicon_trivium_est(&ratum,
-            (i32)LEXICON_C89_GENUS_SPATIA));
+            (s32)LEXICON_C89_GENUS_SPATIA));
         CREDO_FALSUM (materia_lexicon_trivium_est(&ratum,
-            (i32)LEXICON_C89_GENUS_IDENTIFICATOR));
+            (s32)LEXICON_C89_GENUS_IDENTIFICATOR));
 
         /* Genus extra fines: nihil, non ruina */
         CREDO_NIHIL (materia_lexicon_titulus(&ratum,
-            (i32)LEXICON_C89_NUMERUS_GENERUM));
+            (s32)LEXICON_C89_NUMERUS_GENERUM));
         CREDO_AEQUALIS_S32 ((s32)materia_lexicon_species(&ratum,
-            (i32)LEXICON_C89_NUMERUS_GENERUM),
+            (s32)LEXICON_C89_NUMERUS_GENERUM),
             (s32)MATERIA_LEX_NUMERUS_SPECIERUM);
 
         /* DISTRIBUTIO - mensurata 2026-08-27 ante furcam.
          * Numeri pinnati: generatio quae eos movet AUDIRI debet. */
-        per (i = ZEPHYRUM; i < (i32)LEXICON_C89_NUMERUS_GENERUM; i++)
+        per (i = ZEPHYRUM; i < (s32)LEXICON_C89_NUMERUS_GENERUM; i++)
         {
-            commutatio ((s32)materia_lexicon_species(&ratum, i))
+            commutatio ((s32)materia_lexicon_species(&ratum, (s32)i))
             {
             casus (s32)MATERIA_LEX_VERBATIM:   numerus_verbatim++;   frange;
             casus (s32)MATERIA_LEX_FIXUM:      numerus_fixum++;      frange;
@@ -345,7 +345,7 @@ s32 principale (vacuum)
         CREDO_AEQUALIS_I32 (numerus_terminator, (i32)2);
         CREDO_AEQUALIS_I32 (numerus_verbatim + numerus_fixum
                           + numerus_repetitum + numerus_terminator,
-                            (i32)LEXICON_C89_NUMERUS_GENERUM);
+                            (s32)LEXICON_C89_NUMERUS_GENERUM);
     }
